Refactor concurrency control flow in lotus plugin
Refactor management of the notification channel used by workers when they unexpectedly terminate to eliminate
a race condition where the channel could be closed before a worker has written to it. Instead we now rely
on the context to signal that remaining workers should terminate when any other worker has died. This
also eliminates a situation where we deferred, but did not execute, a drain function for every worker run.

Also refactor run to use defer for consistent closure of node api connection and cancelling the worker
context.